Promoted shared JsonViewer component and plugged it into the Network panel too
