+++ packages/replay-next/components/JsonViewer/JsonViewer.module.css
@@ -1,0 +1,8 @@
+.JsonViewer {
+  display: flex;
+  flex-direction: column;
+  font-family: var(--font-family-monospace);
+  font-size: var(--font-size-regular);
+  white-space: pre;
+  padding: 0.5rem 0;
+}
