+++ packages/replay-next/components/JsonViewer/index.ts
@@ -1,0 +1,3 @@
+import { JsonViewer } from "./JsonViewer";
+
+export default JsonViewer;
