+++ packages/replay-next/components/JsonViewer/useJsonViewerContextMenu.tsx
@@ -3,7 +3,7 @@ import { ContextMenuItem, useContextMenu } from "use-context-menu";
 import Icon from "replay-next/components/Icon";
 import { copyToClipboard } from "replay-next/components/sources/utils/clipboard";
 
-export function useRequestDetailsContextMenu(jsonText: string) {
+export function useJsonViewerContextMenu(jsonText: string) {
   const copyJson = () => {
     copyToClipboard(jsonText);
   };
